pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "opencmo"
version = "0.1.0"
description = "Open-source AI CMO multi-agent system for indie developers"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = [
    "openai-agents>=0.12.4",
    "crawl4ai",
    "python-dotenv",
    "httpx>=0.27",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
geo-premium = ["anthropic>=0.49", "google-generativeai>=0.8"]
scheduler = ["apscheduler>=3.11"]
web = ["fastapi>=0.115", "uvicorn>=0.34", "jinja2>=3.1"]
all = ["opencmo[geo-premium,scheduler,web]"]

[project.scripts]
opencmo = "opencmo.cli:main"
opencmo-web = "opencmo.web.app:run_server"
